Remove key-value tests
key-value is used for setting up config, value only for tests.

This removes some of the flexibility but makes it non-ambiguous,
avidining issues where both the key and the description are set at the
same time.

It also makes it possible to have numbered unnamed tests (if only a
function is provided), since they are now always ordered.

Finally, might be beneficial in the future that the test order is
deterministic and obvious.